Fix: auto import amp-facebook-comments
This addresses the problem of optimizer failing to auto import
amp-facebook-comments.

The problem is that the experimental amp-facebook component 1.0 enable
the amp-facebook-comments tag, whereas the current prod version requires
it's own extension.

This fix hardcodes the correct version, but we need to find a generic
way to derive this information from the validator rules.