Determine redirect based on `status` property
Since we were relying on browser spec which supports a
concept of opaque-redirect, the Cloufdlare Workers on the other
hand don't suppport certain properties which make this possible.

They, instead make the actual status code available in redirect
manual mode which is now being used to determine whether the
request is redirect.